fix tela de receitas favoritas
